Users cannot view their workspace, or create new tasks if the subscription for their workspace is expired.

+++ app/Http/Controllers/WorkspaceTasksController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Exceptions\SubscriptionExpiredException;
 use App\Models\Task;
 use App\Models\Workspace;
 use Illuminate\Http\Request;
@@ -17,12 +18,17 @@ class WorkspaceTasksController extends Controller
      */
     public function create(Workspace $workspace)
     {
-        abort_unless($workspace->allMembers()->contains(Auth::user()), 403);
+        try {
+            $this->authorize('update', $workspace);
 
-        return view('tasks.create', [
-            'workspace' => $workspace,
-            'members' => $workspace->allMembers(),
-        ]);
+            return view('tasks.create', [
+                'workspace' => $workspace,
+                'members' => $workspace->allMembers(),
+            ]);    
+        } catch (SubscriptionExpiredException $e) {
+            return response("Subscription exipred. Please renew your subscription.", 423);
+        }
+        
     }
 
     /**
@@ -33,21 +39,28 @@ class WorkspaceTasksController extends Controller
      */
     public function store(Workspace $workspace)
     {
-        $task = Task::create([
-			'created_by' => Auth::user()->id,
-            'workspace_id' => $workspace->id,
-			'name' => request('name'),
-			'user_responsible' => request('user_responsible'),
-			'start_date' => request('start_date'),
-			'finish_date' => request('finish_date'),
-			'status' => request('status'),
-        ]);
+        try {
+            $this->authorize('update', $workspace);
+            
+            $task = Task::create([
+                'created_by' => Auth::user()->id,
+                'workspace_id' => $workspace->id,
+                'name' => request('name'),
+                'user_responsible' => request('user_responsible'),
+                'start_date' => request('start_date'),
+                'finish_date' => request('finish_date'),
+                'status' => request('status'),
+            ]);
 
-        if (request()->wantsJson()) {
-            return response([], 201);
+            if (request()->wantsJson()) {
+                return response([], 201);
+            }
+
+            return redirect(route('workspaces.show', $workspace));
+        } catch (SubscriptionExpiredException $e) {
+            return response("Subscription exipred. Please renew your subscription.", 423);
         }
 
-        return redirect(route('workspaces.show', $workspace));
     }
 
     /**
@@ -58,6 +71,12 @@ class WorkspaceTasksController extends Controller
      */
     public function destroy(Workspace $workspace, Task $task)
     {
-        $task->delete();
+        try {
+            $this->authorize('update', $workspace);
+
+            $task->delete();
+        } catch (SubscriptionExpiredException $e) {
+            return response("Subscription exipred. Please renew your subscription.", 423);
+        }
     }
 }
